Multi Source Cost Distance works and tested. Though think I have found a bug in single cost distance.

+++ spark/src/main/scala/geotrellis/spark/graphx/GraphRDD.scala
@@ -7,5 +7,5 @@ import org.apache.spark.graphx._
 import reflect.ClassTag
 
 class GraphRDD[K: ClassTag](
-  val graph: Graph[(K, Int), Int],
+  val graph: Graph[(K, Double), Double],
   val metaData: RasterMetaData)
